pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "vacuum"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "python-roborock>=4.17.0",
    "mcp>=1.0.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
vacuum = "vacuum.cli:app"
vacuum-mcp = "vacuum.mcp_server:main"

[tool.setuptools.packages.find]
where = ["src"]
